% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My Fitting Linear Models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula class object, which is a symbolic description of
the model to be fitted.}

\item{data}{an input data frame.}
}
\value{
A \code{table} with statistics for this linear model, invluding the
  \code{(Intercept)!} and columns for the \code{Estimate}, \code{Std. Error},
  \code{t value}, and \code{Pr(>|t|)}.
}
\description{
\code{my_lm} is used to fit linear models.
}
\examples{
my_lm(formula = mpg ~ hp + wt, data = mtcars)

}
\keyword{inference}
