% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{My Random Forest Cross-Validation}
\usage{
my_rf_cv(trainx, trainy, k)
}
\arguments{
\item{trainx}{input data frame containing columns of predictor variables.}

\item{trainy}{response vector with length equal to number of rows of
\code{trainx}}

\item{k}{number of cross-validation folds.}
}
\value{
a numeric with the cross-validation error.
}
\description{
This will predict an output using the covariates within the training data
  \code{train}.
}
\keyword{inference}
\keyword{prediction.}
