% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{My t-Test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector input of data}

\item{alternative}{A character string specifying the alternative hypothesis.
It only accepts \code{"two.sided"}, \code{"less"}, or
\code{"greater"}.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
a list with elements \code{test_stat}, the numeric test statistics,
  \code{df}, the degrees of freedom, \code{alternative}, the value of the
  parameter \code{alternative}, and \code{p_val}, the numeric p-value.
}
\description{
This function performs a one sample t-test on vectors of data.
}
\examples{
my_t.test(1:10, alternative = "two.sided", mu = 5)
my_t.test(3:1000, alternative = "less", mu = 100)

}
\keyword{inference}
